refactor(cli): move shared mutation helpers out of restore.go
The split left the create/update/init-shared flag, config-loading and workflow helpers (defaultClusterMutationFieldSelectors, setupMutationCmdFlags, loadAndValidateClusterConfig, runClusterCreationWorkflow, resolveCreatedContextName, and the flag-registration helpers) in restore.go. Moved them to a dedicated mutation.go so restore.go stays focused on restore behavior. Behavior-preserving move within the package.
